Edited landing page with links to About, Products, and Contact page
