+++ pottery/src/app/layout.js
@@ -13,7 +13,7 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-        <NavBar/>
+        {/* <NavBar/> */}
         {children}
       </body>
     </html>
