+++ pottery/src/app/navbar/page.js
@@ -1,17 +1,21 @@
-
+import Link from 'next/link'
 export default async function NavBar() {
 
     return (
-      <div className="navbar">
-        <div id="logo">
-            <h1>Studio Terra</h1>
-        </div>
-        <div className="navLinks">
-            <h1>About</h1>
-            <h1>Products</h1>
-        </div>
- 
-      </div>
+        <header>
+            <div className="navbar">
+                <div id="logo">
+                    <Link href='/'><h3>Studio Terra</h3></Link>
+                </div>
+                <div className="navMenu">
+                    <Link href="/about">About</Link>
+                    <Link href="/products">Products</Link>
+                    <Link href="/contact">Contact</Link>
+                </div>
+        
+            </div>
+        </header>
+      
     )
   }
