+++ pottery/updates.md
@@ -60,3 +60,4 @@ Create an e-commerce for pottery with Next.js, Prisma, and Vercel.
     - Researched on Prisma document ([https://www.prisma.io/docs/orm/overview/prisma-in-your-stack/rest]) and continued to search in this code to see examples of how they retreive data ([https://github.com/prisma/prisma-examples/blob/latest/javascript/rest-nextjs/src/app/page.jsx]). 
     - Finally found a way to fetch data from database without errors. 
 - Creating Navbar
+    - Used this as a refererence: [https://codewithmarish.com/post/how-to-create-responsive-navbar-in-next-js]
